+++ src/model/class.ts
@@ -14,10 +14,12 @@
 
 import * as babel from 'babel-types';
 
+import {ElementMixin} from '..';
 import * as jsdocLib from '../javascript/jsdoc';
-import {Document, ElementMixin, Feature, Method, Privacy, Property, Resolvable, ScannedFeature, ScannedMethod, ScannedProperty, ScannedReference, Severity, SourceRange, Warning} from '../model/model';
+import {Document, Feature, Method, Privacy, Property, Resolvable, ScannedFeature, ScannedMethod, ScannedProperty, ScannedReference, Severity, SourceRange, Warning} from '../model/model';
 import {ParsedDocument} from '../parser/document';
 
+import {DeclaredWithStatement} from './document';
 import {Demo} from './element-base';
 import {ImmutableMap, unsafeAsMutable} from './immutable';
 
@@ -38,6 +40,7 @@ export class ScannedClass implements ScannedFeature, Resolvable {
   /** The name of the class in the local scope where it is defined. */
   readonly localName: string|undefined;
   readonly astNode: babel.Node;
+  readonly statementAst: babel.Statement|undefined;
   readonly jsdoc: jsdocLib.Annotation;
   readonly description: string;
   readonly summary: string;
@@ -54,8 +57,9 @@ export class ScannedClass implements ScannedFeature, Resolvable {
   readonly demos: Demo[];
   constructor(
       className: string|undefined, localClassName: string|undefined,
-      astNode: babel.Node, jsdoc: jsdocLib.Annotation, description: string,
-      sourceRange: SourceRange, properties: Map<string, ScannedProperty>,
+      astNode: babel.Node, statementAst: babel.Statement|undefined,
+      jsdoc: jsdocLib.Annotation, description: string, sourceRange: SourceRange,
+      properties: Map<string, ScannedProperty>,
       methods: Map<string, ScannedMethod>,
       staticMethods: Map<string, ScannedMethod>,
       superClass: ScannedReference<'class'>|undefined,
@@ -64,6 +68,7 @@ export class ScannedClass implements ScannedFeature, Resolvable {
     this.name = className;
     this.localName = localClassName;
     this.astNode = astNode;
+    this.statementAst = statementAst;
     this.jsdoc = jsdoc;
     this.description = description;
     this.sourceRange = sourceRange;
@@ -112,6 +117,7 @@ declare module '../model/queryable' {
 export interface ClassInit {
   readonly sourceRange: SourceRange|undefined;
   readonly astNode: any;
+  readonly statementAst: babel.Statement|undefined;
   readonly warnings?: Warning[];
   readonly summary: string;
   // TODO(rictic): we don't need both name and className here.
@@ -129,11 +135,13 @@ export interface ClassInit {
   readonly privacy: Privacy;
   readonly demos?: Demo[];
 }
-export class Class implements Feature {
+export class Class implements Feature, DeclaredWithStatement {
   readonly kinds = new Set(['class']);
   readonly identifiers = new Set<string>();
   readonly sourceRange: SourceRange|undefined;
+  // TODO(rictic): make this an AstWithLanguage
   readonly astNode: any;
+  readonly statementAst: babel.Statement|undefined;
   readonly warnings: Warning[];
   readonly summary: string;
   readonly name: string|undefined;
@@ -172,6 +180,7 @@ export class Class implements Feature {
       abstract: this.abstract,
       privacy: this.privacy,
       astNode: this.astNode,
+      statementAst: this.statementAst,
       sourceRange: this.sourceRange
     } = init);