+++ src/test/polymer/polymer2-mixin-scanner_test.ts
@@ -467,31 +467,34 @@ Polymer.TestMixin = Polymer.woohoo(function TestMixin(base) {
   test('finds exported mixin functions', async () => {
     const mixins = await getScannedMixins('test-mixin-12.js');
     const mixinData = await Promise.all(mixins.map(getTestProps));
-    assert.deepEqual(mixinData, [{
-                       name: 'TestMixin',
-                       description: 'A mixin description',
-                       summary: '',
-                       properties: [{
-                         name: 'foo',
-                       }],
-                       attributes: [{
-                         name: 'foo',
-                       }],
-                       methods: [],
-                       underlinedWarnings: []
-                     }, {
-                      name: 'DefaultTestMixin',
-                      description: 'Another mixin description',
-                      summary: '',
-                      properties: [{
-                        name: 'bar',
-                      }],
-                      attributes: [{
-                        name: 'bar',
-                      }],
-                      methods: [],
-                      underlinedWarnings: []
-                    });
+    assert.deepEqual(mixinData, [
+      {
+        name: 'TestMixin',
+        description: 'A mixin description',
+        summary: '',
+        properties: [{
+          name: 'foo',
+        }],
+        attributes: [{
+          name: 'foo',
+        }],
+        methods: [],
+        underlinedWarnings: []
+      },
+      {
+        name: 'DefaultTestMixin',
+        description: 'Another mixin description',
+        summary: '',
+        properties: [{
+          name: 'bar',
+        }],
+        attributes: [{
+          name: 'bar',
+        }],
+        methods: [],
+        underlinedWarnings: []
+      }
+    ]);
     const underlinedSource = await underliner.underline(mixins[0].sourceRange);
     assert.deepEqual(underlinedSource, `
 export function TestMixin(superclass) {